+++ pkg/_pub_shared/lib/data/package_api.g.dart
@@ -62,16 +62,21 @@ Map<String, dynamic> _$AutomatedPublishingConfigToJson(
 GithubPublishingConfig _$GithubPublishingConfigFromJson(
         Map<String, dynamic> json) =>
     GithubPublishingConfig(
-      isEnabled: json['isEnabled'] as bool?,
+      isEnabled: json['isEnabled'] as bool? ?? false,
       repository: json['repository'] as String?,
       tagPattern: json['tagPattern'] as String?,
-      requireEnvironment: json['requireEnvironment'] as bool?,
+      requireEnvironment: json['requireEnvironment'] as bool? ?? false,
       environment: json['environment'] as String?,
+      isPushEventEnabled: json['isPushEventEnabled'] as bool? ?? true,
+      isWorkflowDispatchEventEnabled:
+          json['isWorkflowDispatchEventEnabled'] as bool? ?? false,
     );
 
 Map<String, dynamic> _$GithubPublishingConfigToJson(
     GithubPublishingConfig instance) {
-  final val = <String, dynamic>{};
+  final val = <String, dynamic>{
+    'isEnabled': instance.isEnabled,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -79,22 +84,26 @@ Map<String, dynamic> _$GithubPublishingConfigToJson(
     }
   }
 
-  writeNotNull('isEnabled', instance.isEnabled);
   writeNotNull('repository', instance.repository);
   writeNotNull('tagPattern', instance.tagPattern);
-  writeNotNull('requireEnvironment', instance.requireEnvironment);
+  val['requireEnvironment'] = instance.requireEnvironment;
   writeNotNull('environment', instance.environment);
+  val['isPushEventEnabled'] = instance.isPushEventEnabled;
+  val['isWorkflowDispatchEventEnabled'] =
+      instance.isWorkflowDispatchEventEnabled;
   return val;
 }
 
 GcpPublishingConfig _$GcpPublishingConfigFromJson(Map<String, dynamic> json) =>
     GcpPublishingConfig(
-      isEnabled: json['isEnabled'] as bool?,
+      isEnabled: json['isEnabled'] as bool? ?? false,
       serviceAccountEmail: json['serviceAccountEmail'] as String?,
     );
 
 Map<String, dynamic> _$GcpPublishingConfigToJson(GcpPublishingConfig instance) {
-  final val = <String, dynamic>{};
+  final val = <String, dynamic>{
+    'isEnabled': instance.isEnabled,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -102,7 +111,6 @@ Map<String, dynamic> _$GcpPublishingConfigToJson(GcpPublishingConfig instance) {
     }
   }
 
-  writeNotNull('isEnabled', instance.isEnabled);
   writeNotNull('serviceAccountEmail', instance.serviceAccountEmail);
   return val;
 }