+++ CHANGELOG.md
@@ -2,6 +2,9 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2024.05.28`.
+ * Upgraded stable Dart analysis SDK to `3.4.1`
+ * Upgraded stable Flutter analysis SDK to `3.22.1`.
 
  * Upgraded runtime Dart SDK to `3.5.0-164.0.dev`.